+++ examples/ex01_basics/IncrValueModular_test.py
@@ -87,7 +87,7 @@ class IncrTestBench( Component ):
 
 def test_value_modular():
   tb = IncrTestBench()
-  tb.apply( SimulationPass() )
+  tb.apply( DefaultPassGroup() )
 
   # Print out the update block schedule.
   print( "\n==== Schedule ====" )
