+++ examples/ex03_proc/test/ProcFL_test.py
@@ -63,7 +63,7 @@ class ProcFL_Tests:
     th.load( mem_image )
 
     # Create a simulator and run simulation
-    th.apply( SimulationPass(print_line_trace=True) )
+    th.apply( DefaultPassGroup(print_line_trace=True) )
     th.sim_reset()
 
     while not th.done() and th.sim_cycle_count() < max_cycles:
