+++ examples/ex04_xcel/proc-xcel-sim
@@ -219,7 +219,7 @@ def main():
     model.dut.set_metadata( YosysTranslationImportPass.enable, True )
     model = YosysTranslationImportPass()( model )
 
-  model.apply( SimulationPass(print_line_trace=opts.trace) )
+  model.apply( DefaultPassGroup(print_line_trace=opts.trace) )
 
   # Load the program into the model
