+++ pymtl3/__init__.py
@@ -25,7 +25,7 @@ from .dsl.Connectable import (
 from .dsl.ConstraintTypes import RD, WR, M, U
 from .dsl.MetadataKey import MetadataKey
 from .dsl.Placeholder import Placeholder
-from .passes.PassGroups import SimulationPass
+from .passes.PassGroups import DefaultPassGroup
 
 __all__ = [
   'U','M','RD','WR',
@@ -35,7 +35,7 @@ __all__ = [
   'non_blocking', 'CalleeIfcCL', 'CallerIfcCL',
   'blocking', 'CalleeIfcFL', 'CallerIfcFL',
 
-  'SimulationPass',
+  'DefaultPassGroup',
   'Component', 'Placeholder',
 
   'trunc', 'sext', 'zext', 'clog2', 'concat', 'reduce_and', 'reduce_or', 'reduce_xor',
