+++ pymtl3/passes/__init__.py
@@ -1,2 +1,1 @@
 from .PassGroups import *
-from .tracing.TracingConfigs import TracingConfigs
