+++ pymtl3/stdlib/connects/test/connect_bits2bitstruct_test.py
@@ -42,7 +42,7 @@ def test_imatmul_bits2bitstruct_simple():
 
   a = Bits2Bitstruct()
   a.elaborate()
-  a.apply( SimulationPass() )
+  a.apply( DefaultPassGroup() )
   a.pt_bits @= 0xeda
   a.sim_eval_combinational()
   assert a.pt_bitstruct.x == 0xed
@@ -60,7 +60,7 @@ def test_imatmul_bitstruct2bits_simple():
 
   b = Bitstruct2Bits()
   b.elaborate()
-  b.apply( SimulationPass() )
+  b.apply( DefaultPassGroup() )
   b.pt_bitstruct.x @= 0xed
   b.pt_bitstruct.y @= 0xa
   b.sim_eval_combinational()
@@ -78,7 +78,7 @@ def test_imatmul_bits2bitstruct_nested():
 
   a = Bits2BitstructNested()
   a.elaborate()
-  a.apply( SimulationPass() )
+  a.apply( DefaultPassGroup() )
   a.pt_bits @= concat( b6(0x3f), b12(0xeda), b1(1) )
   a.sim_eval_combinational()
   assert a.pt_bitstruct.z    == 0x3f
@@ -98,7 +98,7 @@ def test_imatmul_bitstruct2bits_nested():
 
   b = Bitstruct2BitsNested()
   b.elaborate()
-  b.apply( SimulationPass() )
+  b.apply( DefaultPassGroup() )
   b.pt_bitstruct.z    @= 0x3f
   b.pt_bitstruct.pt.x @= 0xed
   b.pt_bitstruct.pt.y @= 0xa
@@ -154,7 +154,7 @@ def test_connect_bits2bitstruct_simple():
 
   a = Bits2Bitstruct()
   a.elaborate()
-  a.apply( SimulationPass() )
+  a.apply( DefaultPassGroup() )
   a.pt_bits @= 0xeda
   a.sim_eval_combinational()
   assert a.pt_bitstruct.x == 0xed
@@ -170,7 +170,7 @@ def test_connect_bitstruct2bits_simple():
 
   b = Bitstruct2Bits()
   b.elaborate()
-  b.apply( SimulationPass() )
+  b.apply( DefaultPassGroup() )
   b.pt_bitstruct.x @= 0xed
   b.pt_bitstruct.y @= 0xa
   b.sim_eval_combinational()
@@ -186,7 +186,7 @@ def test_connect_bits2bitstruct_nested():
 
   a = Bits2BitstructNested()
   a.elaborate()
-  a.apply( SimulationPass() )
+  a.apply( DefaultPassGroup() )
   a.pt_bits @= concat( b6(0x3f), b12(0xeda), b1(1) )
   a.sim_eval_combinational()
   assert a.pt_bitstruct.z    == 0x3f
@@ -204,7 +204,7 @@ def test_connect_bitstruct2bits_nested():
 
   b = Bitstruct2BitsNested()
   b.elaborate()
-  b.apply( SimulationPass() )
+  b.apply( DefaultPassGroup() )
   b.pt_bitstruct.z    @= b6( 0x3f )
   b.pt_bitstruct.pt.x @= b8( 0xed )
   b.pt_bitstruct.pt.y @= b4( 0xa  )
